BudgetApp in TS - second commit- Budget Field

+++ asset/ts/app.js
@@ -1,13 +1,13 @@
 document.addEventListener("DOMContentLoaded", function (event) {
     initialiseDOM();
     var appController = new Controller();
-    //User Input Field
+    //Variable from User Input Field
     var type, description, value;
     //When click Add button --> Add the Item to the ArrayList
     addBTN.addEventListener('click', function () {
         type = addType.value;
         description = descTXTBOX.value;
-        value = parseInt(valTXTBOX.value);
+        value = parseFloat(valTXTBOX.value);
         appController.addItem(type, description, value);
     });
     //When Remove Button Clicked from Input Field
