+++ asset/ts/app.ts
@@ -3,14 +3,14 @@ document.addEventListener("DOMContentLoaded", function(event) {
     initialiseDOM();
     let appController = new Controller();
     
-    //User Input Field
+    //Variable from User Input Field
     let type, description, value;
     
     //When click Add button --> Add the Item to the ArrayList
     addBTN.addEventListener('click', function(){
         type = addType.value;
         description = descTXTBOX.value;
-        value = parseInt(valTXTBOX.value);
+        value = parseFloat(valTXTBOX.value);
         appController.addItem(type, description, value);
     });
